renaming all client and server under every task
