+++ human_experiments/lab_software/tomcat-baseline-tasks/tasks/ping_pong_task/__init__.py
@@ -1,3 +1,3 @@
-from .client_ai_ping_pong_task import ClientAIPingPongTask
-from .client_ping_pong_task import ClientPingPongTask
-from .server_ping_pong_task import ServerPingPongTask
+from .client_ai import ClientAIPingPongTask
+from .client import ClientPingPongTask
+from .server import ServerPingPongTask
