+++ human_experiments/lab_software/tomcat-baseline-tasks/tasks/ping_pong_task/config.py
@@ -1,8 +1,8 @@
 # count down before start
-SECONDS_COUNT_DOWN = 10.0
+SECONDS_COUNT_DOWN = 2.0
 SHOW_SCORE_COUNT_DOWN_MILLISECONDS = 4000
 
-SESSION_TIME_SECONDS = 120.0
+SESSION_TIME_SECONDS = 5.0
 
 COUNT_DOWN_MESSAGE = "Move the mouse to move the blue paddle"
